Fix blank challenges.
When starting quizzes, look up the questions by words and not by alphagrams. Alphagrams can have blanks or maybe have other forms in the future (for example, they can be BUILD queries or patterns or something of the sort). So we should always look up words instead of alphagrams pretty much when starting quizzes.