Fix date handling for getting challenge LB info.
Note that Django correctly handles dates otherwise! Since USE_L10N is on, when SUBMITTING a challenge request, the date form validator validates the date properly. We should write tests around this behavior, both for the AJAX get (actually, POST :/ ... ) and for the form POST to start a new challenge.